+++ src/halfpipe/tui/templates/selection_template.py
@@ -370,6 +370,27 @@ class SelectionTemplate(Widget):
             rename_item,
         )
 
+    def on_focus_label_selected(self, message: FocusLabel.Selected) -> None:
+        """Changes border title color according to the feature type."""
+
+        # Update content switcher with the newly selected id.
+        # list ids have suffix _flabel, so to match the widget in the content switcher we need to remove this suffix
+        content_switcher_item_new_id = message.control.id[:-7]
+        current_feature = self.feature_items[content_switcher_item_new_id]
+
+        # Use currently selected switcher id to grab the focus label so that we can deselect it.
+        content_switcher_item_old_id = self.get_widget_by_id("content_switcher").current
+        # when we delete item, it can be also a None
+        if content_switcher_item_old_id is not None and content_switcher_item_old_id != content_switcher_item_new_id:
+            collapsible_item_old_id = content_switcher_item_old_id + "_flabel"
+            self.get_widget_by_id(collapsible_item_old_id).deselect()
+
+        self.get_widget_by_id("content_switcher").current = content_switcher_item_new_id
+        self.get_widget_by_id("content_switcher").border_title = "{}: {}".format(
+            self.ITEM_MAP[current_feature.type], current_feature.name
+        )
+        self.get_widget_by_id("content_switcher").styles.border_title_color = "white"
+
     def fill_cache_and_create_new_content_item(self, new_item: tuple[str, str] | list[tuple[str, str]]) -> Widget:
         """
         Fills the cache and creates a new content item.
